move deployment dropwdown to main nav
