+++ src/interfaces/coral_web/src/components/Shared/NavigationBar/NavigationBar.tsx
@@ -3,6 +3,7 @@ import Link from 'next/link';
 import React from 'react';
 
 import { Logo } from '@/components/Shared';
+import { DeploymentsDropdown } from '@/components/Shared/NavigationBar/DeploymentsDropdown';
 
 /**
  * @description Displays the navigation bar where clicking the logo will return the user to the home page.
@@ -21,6 +22,7 @@ export const NavigationBar: React.FC<{ className?: string }> = ({ className = ''
           <Logo />
         </div>
       </Link>
+      <DeploymentsDropdown />
     </nav>
   );
 };
